refactor: 엔드포이트 URL 변경 - 추후 버저닝을 고려하여 RequestMapping에서 v1/~ 이하 부분 제거
