+++ src/main/java/com/bugflix/weblog/like/controller/LikeController.java
@@ -1,6 +1,7 @@
-package com.bugflix.weblog.like;
+package com.bugflix.weblog.like.controller;
 
-import com.bugflix.weblog.like.dto.LikeStatus;
+import com.bugflix.weblog.like.dto.LikeStatusResponse;
+import com.bugflix.weblog.like.service.LikeServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -9,17 +10,15 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.nio.file.Path;
-
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/like")
+@RequestMapping("/api")
 @Slf4j
 public class LikeController {
     private final LikeServiceImpl likeServiceImpl;
 
-    @GetMapping("/{postId}")
-    public ResponseEntity<LikeStatus> changeLikeStatus(@PathVariable(name = "postId") Long postId) {
+    @GetMapping("/v1/like/{postId}")
+    public ResponseEntity<LikeStatusResponse> changeLikeStatus(@PathVariable(name = "postId") Long postId) {
         return ResponseEntity.ok(likeServiceImpl.changeLikeStatus(postId));
     }
 }
