+++ src/main/java/com/bugflix/weblog/page/controller/PageController.java
@@ -1,18 +1,19 @@
-package com.bugflix.weblog.page;
+package com.bugflix.weblog.page.controller;
 
+import com.bugflix.weblog.page.domain.Page;
+import com.bugflix.weblog.page.service.PageServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api/v1/page")
+@RequestMapping("/api")
 @RequiredArgsConstructor
 @Slf4j
 public class PageController {
     private final PageServiceImpl pageService;
 
-    @GetMapping
+    @GetMapping("/v1/page")
     public Page getPage(@RequestParam(name = "url") String url) throws Exception {
 
         return pageService.getPage(url);
