+++ src/main/java/com/bugflix/weblog/post/controller/PostController.java
@@ -1,20 +1,19 @@
-package com.bugflix.weblog.post;
+package com.bugflix.weblog.post.controller;
 
-import com.bugflix.weblog.post.dto.PostPreview;
+import com.bugflix.weblog.post.dto.PostPreviewResponse;
+import com.bugflix.weblog.post.service.PostServiceImpl;
 import com.bugflix.weblog.post.dto.PostResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import com.bugflix.weblog.post.dto.PostRequest;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v1/post")
+@RequestMapping("/api")
 @Slf4j
 public class PostController {
     private final PostServiceImpl postServiceImpl;
@@ -36,7 +35,7 @@ public class PostController {
      * - 1. extension 에서 저장 요청
      * - 2. weblog page 에서 저장 요청
      */
-    @PostMapping
+    @PostMapping("/v1/post")
     public ResponseEntity<Void> savePost(@RequestBody PostRequest postRequest) {
         postServiceImpl.savePost(postRequest);
         return ResponseEntity.ok().build();
@@ -59,7 +58,7 @@ public class PostController {
      * - 1. extension 에서 수정 요청
      * - 2. weblog page 에서 수정 요청
      */
-    @PutMapping
+    @PutMapping("/v1/post")
     public ResponseEntity<Void> updatePost(@RequestBody PostRequest postRequest, @RequestParam(name = "postId") Long postId) throws Exception {
         postServiceImpl.updatePost(postRequest, postId);
         return ResponseEntity.ok().build();
@@ -82,7 +81,7 @@ public class PostController {
      * - 1. post 목록에서 post Click
      * - 2. post_id 로 post 검색
      */
-    @GetMapping("/{postId}")
+    @GetMapping("/v1/post/{postId}")
     public ResponseEntity<PostResponse> getPost(@PathVariable Long postId) throws Exception {
 
         return ResponseEntity.ok(postServiceImpl.getPost(postId));
@@ -104,7 +103,7 @@ public class PostController {
      * Scenario :
      * - 1. extension 에서 post 보기 요청
      */
-    @GetMapping
+    @GetMapping("/v1/post")
     public ResponseEntity<List<PostResponse>> getPost(@RequestParam(name = "url") String url) {
 
         return ResponseEntity.ok(postServiceImpl.getPosts(url));
@@ -115,7 +114,7 @@ public class PostController {
      * Parameter :
      * - String url: web page 주소
      * Return :
-     * - ResponseEntity<ArrayList<PostPreview>> :
+     * - ResponseEntity<ArrayList<PostPreviewResponse>> :
      * - Success : 200 ok && ArrayList<Post>
      * - Failed : error
      * <p>
@@ -125,8 +124,8 @@ public class PostController {
      * Scenario :
      * - 1. extension 에서 post 목록 요청
      */
-    @GetMapping("/preview")
-    public ResponseEntity<List<PostPreview>> getPostPreview(@RequestParam(name = "url") String url) {
+    @GetMapping("/v1/post/preview")
+    public ResponseEntity<List<PostPreviewResponse>> getPostPreview(@RequestParam(name = "url") String url) {
 
         return ResponseEntity.ok(postServiceImpl.getPostPreview(url));
     }
@@ -136,13 +135,13 @@ public class PostController {
      * Parameter :
      * - String url
      * Return :
-     * - ResponseEntity<ArrayList<PostPreview>>
+     * - ResponseEntity<ArrayList<PostPreviewResponse>>
      * <p>
      * Explanation :
      * - 특정 web page 에 있는 본인이 작성한 모든 post 의 preview 반환
      */
-    @GetMapping("/mine")
-    public ResponseEntity<List<PostPreview>> getMyPostPreview(@RequestParam(name = "url") String url) {
+    @GetMapping("/v1/post/mine")
+    public ResponseEntity<List<PostPreviewResponse>> getMyPostPreview(@RequestParam(name = "url") String url) {
 
         return ResponseEntity.ok(postServiceImpl.getMyPostPreview(url));
     }
@@ -157,7 +156,7 @@ public class PostController {
      * Explanation :
      * - postId 로 Post 삭제
      */
-    @DeleteMapping("/{postId}")
+    @DeleteMapping("/v1/post/{postId}")
     public ResponseEntity<Void> deletePost(@PathVariable Long postId) {
 
         postServiceImpl.deletePost(postId);
